fix: migration fixed to include role management
