+++ AcademicProgrammingApplication/models/__init__.py
@@ -7,3 +7,4 @@ from .Teacher import Teacher
 from .Viatic import Viatic
 from .User import User
 from .Role import Role
+from .PlanningProposal import PlanningProposal
